Move last parts to healtsites
Migrations and models left in event_mapper
