--- django_project/event_mapper/tasks/__init__.py
@@ -1,4 +1,0 @@
-# -*- coding: utf-8 -*-
-from __future__ import unicode_literals, print_function, absolute_import, division
-
-from .daily_pdf_report import daily_pdf_report
