+++ django_project/event_mapper/urls.py
@@ -1,0 +1,9 @@
+# -*- coding: utf-8 -*-
+from __future__ import absolute_import, division, print_function, unicode_literals
+
+from django.conf.urls import patterns, url
+
+urlpatterns = patterns(
+    '',
+
+)
