+++ django_project/healthsites/data/ne_10m_admin_0_countries.dbf
